feat(files): add BASE_FILE_PATH constant and refactor path handling
